requirements.txt
opencv-contrib-python>=4.8.1.78
mediapipe>=0.10.9
numpy>=1.26.0
tensorflow>=2.16.1
pyopengl
pygame
pyaudio
fastapi
uvicorn
python-dotenv
fastapi
uvicorn
websockets